refactor: improve typings
