+++ src/services/shared-resource-loader.ts
@@ -1,16 +1,11 @@
-import type {
-  createSharedResource,
-  SharedResourceEntry,
-} from '../modules/helpers/shared-resources.js';
+import type { SharedResourceEntry } from '../@types/shared-resources.js';
+import type { createSharedResource } from '../modules/helpers/shared-resources.js';
 import { relative } from 'node:path';
 import { pathToFileURL } from 'node:url';
 import { GLOBAL } from '../configs/poku.js';
 import { isWindows } from '../parsers/os.js';
 
-/**
- * Execute resource files (*.resource.ts) in the current process
- * to initialize shared resources before running tests
- */
+/** Execute resource files (*.resource.ts) in the current process to initialize shared resources before running tests. */
 export async function executeResourceFiles(
   files: string[],
   registry: Record<string, SharedResourceEntry>,
@@ -21,17 +16,16 @@ export async function executeResourceFiles(
 ): Promise<void> {
   for (const file of files) {
     const { entry, name, cleanup } = await executeResourceFile(file);
-    registry[name] = entry as SharedResourceEntry;
+
+    registry[name] = entry;
+
     if (cleanup) {
       cleanupMethods[name] = cleanup;
     }
   }
 }
 
-/**
- * Execute a single resource file in the parent process
- * Loads the resource module, gets the default export (should be { entry, name }), and registers it
- */
+/** Execute a single resource file in the parent process loads the resource module, gets the default export (should be { entry, name }), and registers it. */
 export async function executeResourceFile(path: string) {
   const { cwd } = GLOBAL;
   const file = relative(cwd, path);
@@ -57,9 +51,7 @@ export async function executeResourceFile(path: string) {
   return resource;
 }
 
-/**
- * Filter out resource files from the test files list
- */
+/** Filter out resource files from the test files list */
 export function separateResourceFiles(files: string[]): {
   resourceFiles: string[];
   testFiles: string[];
